Prototype: Basic take movement working
? Spent quite a bit of time fighting with antialiasing because I (think 
I) want the janky pixelly nature of Combat. However even with it off the 
way the tank rotates doesn't look like the way the Atari tank rotates, 
so it may be I have to build individual frames for every rotation to 
match it properly.

? Movement stuff is way too smooth with default settings so I slowed 
down the response of the controls, though it now occurs to me I could 
possible have just turned down the framerate? But no, because the 
rotation of the tank on death in the original game is super fast right? 
So probably I'm doing it the right way.

? All this is in the name of authenticity to the original form, in 
particular because I'd like to do all the work at the original 
resolution (or as close as I can get it without too much work). I'm a 
little worried about things like displaying text at a tiny little 
resolution, but I guess there are fonts that can do this? I mean, Atari 
does it for one. Better not to rely overly on text for that reason? Like 
Atari games.

? Maybe each game can have a manual page that can be triggered somehow 
with a click or something, which could pop up in a separate window (to 
avoid resolution issues with displaying it in the game itself)? Might be 
fun to work out how to communicate the rules of the game(s) this way? 
Perhaps just a manual for the whole game actually?

? In keeping with the ultra-low resolution, I'll have to be very, very 
cautious with representations. Nothing complex is going to work in the 
slightest!