Prototype: pixel collision hell
? After too much work I have a version of almost pixel-perfect collision 
working, in keeping with the original game. I don't know that it was 
worth it and it's still pretty janky - though I observe the original 
game also has pretty nutty collision stuff that goes on too, so perhaps 
I don't mind.

? Currently I don't react to overlaps that occur when rotating without 
speed, whereas Combat does react (violently!) to that situation, so I 
probably ought to as well.

? I think this is closer to a "good enough" state, where I feel like the 
behaviour is sufficiently related that I don't need to feel too bad?