The Conversation: created a basic version
? Spent time trying to make it be more literal to the movie with the 
hotel scene, but eventually have felt it's better to go the more 
"conceptual" route where the role of the player is to listen "through 
the wall" of a standard combat arena. It's even interesting that there's 
that space above the arena in the standard game that allows a spot for 
the player to lurk and listen. And that combines well with the invisible 
variants of the game which create a precedent for not being able to see 
the tanks as they move around.

? Will require significantly better AI to make sure the tanks eventually 
find each other and one shoots (murders) the other. Interesting to frame 
it as a murder (maybe only one tank should be able to shoot? No.)

? The soundscale is suitably disorienting and also proper. I think this 
one is a good solution.