AI Tank: basic AI continues with shooting and random rotation
? It's better and maybe it's good enough? I mean, a human play will 
obviously destroy this AI easily, but at least it kind of mucks around 
and just bother to start shooting when in some kind of range.

? It may well be that this is sufficient given how little any of the 
levels actually rely on an AI. Really it's The Conversation which is the 
big one, and there you can't even see what's happening so it may be 
totally fine. Probably not going to be worth writing substantially more 
AI stuff I suspect.