The Godfather: added instruction, endings, and reduced enemies to one
? Reduced enemies to respect the sprite limits of the Atari system. It's 
a real tension between evoking the scene itself which is very much about 
a multitude of gunmen all killing Sonny together, versus the constraints 
of Combat as a game, which doesn't allow for multiple enemies in that 
way. There's no correct resolution, but I feel "safer" respecting the 
hardware requirement for some reason. The solo killer still captures the 
helplessness of sonny as he's murdered by surprise?

? It's called a "toll plaza" apparently, so I used that terminology. 
Similarly, it's Barzini's people that kill Sonny, so I included that 
too. Little specifics from the language of the movie help to anchor 
things further I imagine...