Au Hasard Balthazar: added timer, endings
? Well it works, though I suspect there's slightly better language 
available for the various parts. The instruction doesn't feel like an 
instruction, but then the movie is kind of instructionless in a way? 
Maybe it should be about making a choice whether to be cruel or kind per 
the people in the movie? Or indifferent? You can't really be kind?

? It's (still) fun that Balthazar moves essentially based on random 
behaviour (au hasard), when he's not avoiding walls. Should he have 
purely random-based behaviour to match the title of the film completely? 
Although in a sense it is still random based all the way down... but 
perhaps avoiding walls is too intentional?