2001: A Space Odyssey: A more working version
? Added instructions and game overs for clarity, which is fine and 
dandy.

? Added an invisible wall that stops the enemy tank from crossing toward 
you and thus potentially touching the monolith which it shouldn't do. 
Thought for some time about having this invisible wall be literally 
represented as a river, then realized it's a "watering hole" in the 
film/book, then realized that getting too literally with the scene was 
feeling pretty off, so went with the more metaphorical approach to the 
monolith as the giver of shooting/death.

? Some truly awful hard-coding of numbers present at the moment in part 
because you can't get the "center" of a set of tiles like the monolith, 
almost worth positioning a sprite in the same location or literally just 
using a sprite separate from the tile map? Why didn't I think of that 
already? Fuck.

? It basically works and basically makes sense.

? I fear there are looming bugs because of the crappy way I handle 
shooting now that I need to actually do something specific on 
player/enemy death. Makes me think I should invest in understanding some 
custom event dispatching, would probably be good to know in general.