Citizen Kane: instructions, multiple memories
? Well, I've added a first attempt at a Combat-style instructions manual 
thing for the start of Citizen Kane. I think it definitely improves 
matters and allows a lot more flavour to be added that sets up the game 
in a fun way. I think I especially appreciate the parallel between the 
splash page/box art idea of representing something really complex in a 
really simple form, too - the text allows you to build up all this 
representational work in imagination, only to have it be (kind of) let 
down by the game itself in a fun way?