The Conversation: removed the player tank, renamed enemy tanks
? Realized that a) having the player sprite plus two enemy sprites 
breaks Atari sprite rules, and b) although there's a player tank 
everywhere else, all the player actually does in this game is listen, so 
the tank is extraneous (completely obvious when you consider I placed it 
outside the playing field). With it gone the game is much more pure and 
accurate.

? Renamed the two AI tanks to husband and wife to keep things at the 
variable name level consistent with the movie (and also with the 
instructions, which reference a "husband tank and wife tank". I like 
this kind of thing, regardless of whether anyone ever sees it.

? Also rewrote the instructions accordingly