Multiple changes!
? The Conversation tanks now point in different directions at the 
beginning and there are no walls to encourage a faster but not 
instantaenous conflict between them, as well as the sound of shooting 
earlier in the piece to simulate hearing the conflict early on.

? The Conversation tank that is killed becomes visible during its death 
spin to reference the visibility of the murdered person in the frosted 
window. The player doesn't know which tank is which (unlike the code) 
and so it's a mysterious sudden vision of murder without knowing what 
happened, fitting the movie relatively well!

? Also fixed various issues of missing death spins and incorrect endings 
in a couple of places.